+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# Changelog for Cldr_Numbers v2.18.0
+
+This is the changelog for Cldr v2.18.0 released on May 8th, 2021.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-cldr/cldr_numbers/tags)
+
+### Big Fixes
+
+* Fixes formatting whent the format string consists only of digits. Previously this would erroneously set both the maximum and minimum integer digits. Now it only sets the minimum integer digits.
+
+### Enhancements
+
+* Add option `:maximum_integer_digits` as an option to `Cldr.Number.to_string/2`
+
 # Changelog for Cldr_Numbers v2.17.0
 
 This is the changelog for Cldr v2.17.0 released on April 8th, 2021.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-cldr/cldr_numbers/tags)